refactor: creates separate http client file for server
