+++ src/tests/utils.ts
@@ -1,5 +1,4 @@
-import { HttpClient } from '@atb/modules/api-server';
-import { HttpEndpoints } from '@atb/modules/api-server/utils';
+import type { HttpClient, HttpEndpoints } from '@atb/modules/api-server';
 import type {
   GetServerSideProps,
   GetServerSidePropsContext,
